refactor: split huggingface models and adaptors into separate modules
