+++ peekingduck/pipeline/nodes/model/huggingface_hubv1/adaptors/__init__.py
@@ -1,0 +1,20 @@
+# Copyright 2022 AI Singapore
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Hugging Face Hub model adaptors."""
+
+from .base import HuggingFaceAdaptor
+from .instance_segmenter import InstanceSegmenter
+from .object_detector import ObjectDetector
+from .panoptic_segmenter import PanopticSegmenter
